Respond to CR.
// FREEBIE
